requirements.txt
flask==2.3.3
torch>=1.12.0
transformers>=4.25.0
deepmultilingualpunctuation==1.0.1
numpy>=1.23.0
scipy>=1.10.0
regex
language-tool-python==2.7.1
pyspellchecker==0.7.2
